tohep: rename crawler2hep module to tohep
Now it contains more than just crawler format to hep format
converters.
