+++ tests/unit/responses/tohep/in_no_document_type.yaml
@@ -1,9 +1,11 @@
+# Simple document that should fallback to 'article' document type upon hep
+# record generation.
 {
     "page_nr": [
         "11"
     ],
     "acquisition_source": {
-        "date": "2017-02-21T18:03:40.858985",
+        "datetime": "2017-02-21T18:03:40.858985",
         "source": "arXiv",
         "method": "hepcrawl",
         "submission_number": "scrapy_job"
@@ -46,15 +48,6 @@
         },
         {
             "primary": "Review"
-        },
-        {
-            "primary": "Book"
-        },
-        {
-            "primary": "Bookchapter"
-        },
-        {
-            "primary": "Conferencepaper"
         },
         {
             "primary": "Citeable"
