+++ tests/unit/responses/tohep/out_no_document_type.yaml
